requirements.txt
requests
pillow
pandas
tqdm
